+++ pkg/microservice/aslan/core/common/repository/models/project_management.go
@@ -1,0 +1,32 @@
+/*
+ * Copyright 2022 The KodeRover Authors.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package models
+
+import "go.mongodb.org/mongo-driver/bson/primitive"
+
+type ProjectManagement struct {
+	ID        primitive.ObjectID `bson:"_id,omitempty" json:"id" `
+	Type      string             `bson:"type"         json:"type"`
+	JiraHost  string             `bson:"jira_host"    json:"jira_host"`
+	JiraUser  string             `bson:"jira_user"    json:"jira_user"`
+	JiraToken string             `bson:"jira_token"   json:"jira_token"`
+	UpdatedAt int64              `bson:"updated_at"   json:"updated_at"`
+}
+
+func (ProjectManagement) TableName() string {
+	return "project_management"
+}
