+++ pkg/microservice/aslan/server/rest/router.go
@@ -48,7 +48,6 @@ import (
 	connectorHandler "github.com/koderover/zadig/pkg/microservice/systemconfig/core/connector/handler"
 	emailHandler "github.com/koderover/zadig/pkg/microservice/systemconfig/core/email/handler"
 	featuresHandler "github.com/koderover/zadig/pkg/microservice/systemconfig/core/features/handler"
-	jiraHandler "github.com/koderover/zadig/pkg/microservice/systemconfig/core/project_management/handler"
 	userHandler "github.com/koderover/zadig/pkg/microservice/user/core/handler"
 
 	// Note: have to load docs for swagger to work. See https://blog.csdn.net/weixin_43249914/article/details/103035711
@@ -121,7 +120,6 @@ func (s *engine) injectRouterGroup(router *gin.RouterGroup) {
 	for _, r := range []injector{
 		new(connectorHandler.Router),
 		new(emailHandler.Router),
-		new(jiraHandler.Router),
 		new(configcodehostHandler.Router),
 		new(featuresHandler.Router),
 	} {
